requirements.txt
streamlit==1.39.0
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
groq==0.4.1
httpx==0.27.2
python-dotenv==1.0.0
rich==13.9.4
